# Parallelize refundedGas for big transactions