Parallelizing postRetcon and refundedGas improved the performances by around 10 % 

**Before this PR**

- Number of blocks : 97, each block is 7 million gas
- Mean: 1092.57 ms
- Median: 1093.0 ms
- 75th Percentile: 1128.0 ms
- 95th Percentile: 1182.0 ms
- 99th Percentile: 1188.75 ms
<img width="1716" alt="image" src="https://github.com/Consensys/linea-arithmetization/assets/5099602/0eff9196-529d-4a38-92c5-3bfe60e91cd2">


**With this PR**
- Number of blocks ': 99, each block is 7 million gas
- Mean: 993.59 ms
- Median: 951.0 ms
- 75th Percentile: 1082.5 ms
- 95th Percentile: 1220.3 ms
- 99th Percentile: 1260.45 ms
<img width="1716" alt="image" src="https://github.com/Consensys/linea-arithmetization/assets/5099602/c448dbe4-0701-4f10-b428-38caf7dea7cb">

Fixes https://github.com/Consensys/linea-arithmetization/issues/792